DB 구조 개선
막상 kr.lang 파일을 위해서 DB 조회할 생각을 하니 문장의 키마다 하나씩 있는 것이 좋다고 생각을 했습니다.

+++ WebProject/database/migrations/2024_07_26_193812_create_translation_logs_table.php
@@ -20,7 +20,7 @@ class CreateTranslationLogsTable extends Migration
             $table->mediumInteger('index');        // lang 파일 Index 열
             $table->integer('offset');             // lang 파일 Offset 열
             $table->text('text');                  // lang 파일 Text 열
-            $table->tinyInteger('version');        // 이 레코드를 넣게된 버전
+            $table->smallInteger('version');       // 이 레코드를 넣게된 버전
             $table->tinyInteger('state');          // 번역 상태
             $table->tinyInteger('user_id');        // 번역자
             $table->timestamps();
